+++ lang/en/local_navbarplus.php
@@ -27,11 +27,20 @@ defined('MOODLE_INTERNAL') || die();
 $string['pluginname'] = 'Navbar Plus';
 $string['setting_inserticonswithlinks'] = 'Icons with links';
 $string['setting_inserticonswithlinks_desc'] = 'With this setting you can add link icons to the header navbar left to the icons "messages" and "notifications".<br/>
-Each line consists of an icon image, a link URL and a text, separated by pipe characters. Each icon needs to be written in a new line. <br/>
+Each line consists of an icon image, a link URL, a text, supported language(s) (optional) and new window setting (optional) - separated by pipe characters. Each icon needs to be written in a new line.<br/>
 For example:<br/>
-a/help|http://moodle.org|Moodle <br/>
-fa-sign-out|/login/logout.php|Logout <br/><br/>
-Further information to the parameters:<br/>
-<b>Image:</b> You can add identifiers for a Moodle icon from the pix folder (<a href="https://github.com/moodle/moodle/tree/master/pix">see the icon list on github.com</a>) or a Font Awesome icon identifier (<a href="http://fontawesome.io/icons/">See the icon list on fontawesome.io</a>). Font Awesome is not included in Boost from Moodle core, but there are some Boost child themes and other solutions from third parties which add Font Awesome support to Moodle.<br/>
-<b>Link:</b> The link target can be defined by a full web URL (e.g. https://moodle.org) or a relative path within your Moodle instance (e.g. /login/logout.php). <br/>
-<b>Title:</b> This text will be written in the title and alt attributes of the icon.';
+a/help|http://moodle.org|Moodle|en,de|true<br/>
+fa-sign-out|/login/logout.php|Logout||false<br/><br/>
+Further information to the parameters:
+<ul>
+<li><b>Image:</b> You can add identifiers for a Moodle icon from the pix folder (<a href="https://github.com/moodle/moodle/tree/master/pix">see the icon list on github.com</a>) or a Font Awesome icon identifier (<a href="http://fontawesome.io/icons/">See the icon list on fontawesome.io</a>). Font Awesome is not included in Boost from Moodle core, but there are some Boost child themes and other solutions from third parties which add Font Awesome support to Moodle.</li>
+<li><b>Link:</b> The link target can be defined by a full web URL (e.g. https://moodle.org) or a relative path within your Moodle instance (e.g. /login/logout.php). </li>
+<li><b>Title:</b> This text will be written in the title and alt attributes of the icon.</li>
+<li><b>Supported language(s) (optional):</b> This setting can be used for displaying the link to users of the specified language only. Separate more than one supported language with commas. If the link should be displayed in all languages, then leave this field empty.</li>
+<li><b>New window (optional)</b>:  By default the link will be opened in the same window and the value of this setting is set to false. If you want to open the link in a new window set the value to true.</li>
+</ul>
+Please note:
+<ul>
+<li> Pipe dividing for optional parameters is always needed if they are located between other options. This means that you have to separate params with the pipe character although they are empty. Also see the example for the Font Awesome icon above. </li>
+<li> If the icon does not show up in the navbar, please check if all mandatory params are set correctly and if the optional language setting fits to your current Moodle user language. </li>
+</ul>';